requirements.txt
python-dotenv
google-genai
pillow
pyinstaller